fix: fix the position of signup file
